Add darcy_friction_haaland function and update first wall model to use it
